kr3/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic[email]>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
slowapi>=0.1.9
python-multipart>=0.0.6
